Move `communicate` to another Rust module
